Генерация C интерфейсов для App

+++ crates/app-base/examples/app-base-example1.rs
@@ -28,7 +28,7 @@ const MAX_ITERS: usize = 100_000;
 const FILE_PATH: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/examples/data.ini");
 
 #[unsafe(no_mangle)]
-fn main(argc: usize, argv: *const *const c_char) -> c_int {
+fn main(argc: c_int, argv: *const *const c_char) -> c_int {
     dotenv(false);
     log_init();
