+++ crates/app/examples/app-example1.rs
@@ -20,7 +20,7 @@ fn main() -> Void {
 
 #[cfg(not(feature = "std"))]
 #[unsafe(no_mangle)]
-fn main(argc: usize, argv: *const *const c_char) -> c_int {
+fn main(argc: c_int, argv: *const *const c_char) -> c_int {
     let _app = App::boot(argc, argv).unwrap();
     libc::EXIT_SUCCESS
 }
